+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.2.0 (2022-03-22)
 ### Packaging
 - Drop support for Python 3.6. [#209](https://github.com/PyO3/setuptools-rust/pull/209)